refactor(clawdstrike-policy-event): extract kinds/deception_{materialization,cleanup,rotation} from edr/receipt/mod.rs
